emailing_system/requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
idna==2.5
mws==0.7.3
python-dateutil==2.6.1
requests==2.18.2
schedule==0.4.3
six==1.10.0
urllib3==1.22
